controller factura

+++ Problema1_5Repositorio/Services/FacturaService.cs
@@ -13,24 +13,43 @@ namespace Problema1_5.Services
     {
         private readonly IFacturaRepository _facturaRepository;
 
-      public FacturaService()
+        public FacturaService()
         {
             _facturaRepository = new FacturaRepository();
         }
         public List<Factura> GetAll()
         {
-            return _facturaRepository.GetAll(); 
+            return _facturaRepository.GetAll();
         }
         public Factura GetById(int id)
         {
             return _facturaRepository.GetById(id);
         }
+        public List<Factura> GetByPaymentOrDate(DateTime? date, int? payment)
+        {
+            return _facturaRepository.GetAll().Where(f => f.Fecha == date || f.FormaPagoId == payment).ToList();
+
+        }
         public bool Save(Factura cliente)
         {
             return _facturaRepository.Save(cliente);
         }
+        public bool Update(Factura factura)
+        {
+            var articuloExists = GetById(factura.Id);
+            if (articuloExists == null)
+            {
+                return false;
+            }
+            return _facturaRepository.Update(factura);
+        }
         public bool Delete(int id)
         {
+            var facturaExists = GetById(id);
+            if (facturaExists == null)
+            {
+                return false;
+            }
             return _facturaRepository.Delete(id);
 
         }
